requirements.txt
Flask==2.3.3
Werkzeug==2.3.8
Jinja2==3.1.2
itsdangerous==2.1.2
pillow==10.3.0
python-magic==0.4.27
pytest==8.3.3
waitress==3.0.0
